Remove researchers section
I am removing researchers section, it got slipped in but I don't think it's a great idea. Either anyone who has ever published any paper or exploit or reported a vulnerability can be listed or someone has to make a call on whom to list. The former makes the list unwieldy and of little use. The latter makes the list subjective. In either case, the main goal of this README is to primarily list useful work relevant to Android security and not individuals who are doing it.